refactor: move z4jspec out of client dir
