+++ src/test/groovy/lol/pbu/z4j/client/ArticlesClientSpec.groovy
@@ -2,6 +2,7 @@ package lol.pbu.z4j.client
 
 import io.micronaut.http.HttpStatus
 import io.micronaut.test.extensions.spock.annotation.MicronautTest
+import lol.pbu.z4j.Z4jSpec
 import spock.lang.Shared
 import spock.lang.Unroll
