+++ src/test/groovy/lol/pbu/z4j/client/UserSegmentsClientSpec.groovy
@@ -2,6 +2,7 @@ package lol.pbu.z4j.client
 
 import io.micronaut.http.HttpStatus
 import io.micronaut.test.extensions.spock.annotation.MicronautTest
+import lol.pbu.z4j.Z4jSpec
 import lol.pbu.z4j.model.CreateUserSegmentRequest
 import lol.pbu.z4j.model.UserSegment
 import spock.lang.Shared
